[FEATURE] Adapted events page and styles
