+++ desparchado/frontend/scripts/event-details.ts
@@ -1,0 +1,6 @@
+import '../styles/pages/event-detail.scss';
+import { attachOnLoadListener } from './utils/page-load-listener';
+
+const registry = new Map();
+
+attachOnLoadListener(registry);
