+++ desparchado/frontend/scripts/events.ts
@@ -1,0 +1,6 @@
+import '../styles/pages/events.scss';
+import { attachOnLoadListener } from './utils/page-load-listener';
+
+const registry = new Map();
+
+attachOnLoadListener(registry);
