+++ desparchado/frontend/styles/forms/inputs/select-group.scss
@@ -1,0 +1,74 @@
+@use '@styles/variables' as *;
+@use '@styles/breakpoints' as *;
+@use '@styles/functions' as *;
+@use '@styles/typography' as *;
+@use '@styles/animations' as *;
+@use '@styles/mixins' as *;
+@use "sass:map";
+
+.select-group {
+  background: $color-dp-orange-100;
+  border: $hairline_regular;
+  border-color: $color-layout-interactive-background;
+  border-radius: 100rem;
+  width: fit-content;
+  display: flex;
+
+  &__item {
+    @extend .text-body-highlight;
+
+    border-radius: 100rem;
+    position: relative;
+
+    &::after {
+      content: '';
+      position: absolute;
+      top: 0;
+      left: 0;
+      width: 100%;
+      height: 100%;
+      background: $color-black-100;
+      border-radius: 100rem;
+      opacity: 0;
+      transition: 0.5s ease opacity;
+      pointer-events: none;
+    }
+
+    &:hover {
+      &::after {
+        opacity: 1;
+      }
+    }
+
+    select {
+      font-weight: $f-weight-semibold;
+      background: none;
+      border: none;
+      padding: 0.75rem 1.5rem;
+      padding-right: 2rem;
+      -moz-appearance:none;
+      -webkit-appearance:none;
+      appearance:none;
+      border-radius: 100rem;
+      cursor: pointer;
+      min-height: 2.75rem;
+
+      &:focus-visible {
+        outline: 0.125rem solid $color-dp-orange-300;
+        outline-offset: 0.125rem;
+      }
+    }
+
+    &--emphasis {
+      background: $color-layout-interactive-background;
+
+      select {
+        color: $color-layout-background;
+
+        option {
+          color: $color-layout-foreground;
+        }
+      }
+    }
+  }
+}
