+++ desparchado/frontend/styles/pages/event-detail.scss
@@ -1,3 +1,3 @@
 @use '@styles/sections/intro-section';
-@use '@styles/sections/events-list-section';
+@use '@styles/sections/event-details';
 @use '@styles/sections/footer';
