+++ desparchado/frontend/styles/pages/events.scss
@@ -1,3 +1,4 @@
 @use '@styles/sections/intro-section';
-@use '@styles/sections/events-list-section';
+@use '@styles/sections/events-section';
+@use '@styles/forms';
 @use '@styles/sections/footer';
