+++ events/templates/events/event_detail.html
@@ -37,7 +37,7 @@
 {% endblock %}
 
 {% block extra_scripts %}
-  {% vite_asset 'desparchado/frontend/scripts/home.ts' %}
+  {% vite_asset 'desparchado/frontend/scripts/event-details.ts' %}
 
   <script
       src="https://maps.googleapis.com/maps/api/js?key={{ GOOGLE_MAPS_API_KEY }}&libraries=maps,marker"
